chore: Regenerate `yarn.lock` in the hopes of fixing a weird CI bug
Nothing was changed about the lock file in the previous commits and it had worked with the same dependencies before => I'm not sure if this will fix it, but it's the next logical step in trying to fix this.

This also includes an adjustment of the `package.json` that was necessary to get it to work locally (although the local problem was a different one than the one in the CI workflow).